refactor(template)!: delete v1 TextRenderer, HCL2 is the only engine (C.6)
Phase C.6 of IMPL-0004: with the validators rejecting v1 and the
--experimental-hcl2 flag gone, the v1 text/template renderer has no
production callers. Delete it and collapse the package down to a
single Renderer.

Removed:
  * internal/template/cty.go — ToCtyValues/FromCtyValues here was the
    v1-transition adapter. Typed lockfile.ToCtyValues/FromCtyValues is
    the canonical conversion now (introduced in C.4).
  * Old internal/template/renderer.go (TextRenderer, NewTextRenderer,
    NewRenderer alias, RenderFile/RenderString/RenderPath/EvaluateBool
    on text/template, normalizePathTemplate shorthand normalisation).
  * Old internal/template/funcs.go FuncMap and v1 helper wrappers
    (replace/trimPrefix/trimSuffix/defaultVal). HCL2 uses
    cty/function/stdlib for those.
  * internal/template/funcs_test.go.
  * internal/template/hcl_*_test.go (renamed below).

Renamed & consolidated:
  * hcl_renderer.go → renderer.go. HCLRenderer struct renamed to
    unexported hclRenderer; NewHCLRenderer renamed to NewRenderer.
    The Renderer interface is the single public surface. The package
    keeps StripTemplateExtension and IsTemplate.
  * hcl_funcs.go → funcs.go. Pure-Go case helpers (snakeCase,
    camelCase, pascalCase, kebabCase, splitWords, toDelimited,
    capitalize) live here alongside their cty.Function wrappers; the
    custom functions still call the Go helpers internally so there is
    no duplicated case logic.

Callers updated:
  * internal/check/check.go, internal/sync/engine.go,
    internal/create/create.go, internal/create/conditions_test.go,
    internal/migratecmd/integration_test.go: NewHCLRenderer →
    NewRenderer.
