+++ internal/sync/engine.go
@@ -54,7 +54,7 @@ func Run(opts *Opts) (*Result, error) {
 	}
 
 	result := &Result{}
-	renderer := tmpl.NewHCLRenderer()
+	renderer := tmpl.NewRenderer()
 
 	bpVars := loadBlueprintVariables(opts.RegistryDir, lock.Blueprint.Path)
